+++ sync-for-expenses/RELEASES.md
@@ -689,3 +689,13 @@ Based on:
 - [typescript v3.1.0] sync-for-expenses
 ### Releases
 - [NPM v3.1.0] https://www.npmjs.com/package/@codat/sync-for-expenses/v/3.1.0 - sync-for-expenses
+
+## 2023-10-05 00:16:26
+### Changes
+Based on:
+- OpenAPI Doc prealpha https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Sync-Expenses.yaml
+- Speakeasy CLI 1.94.0 (2.147.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.1.1] sync-for-expenses
+### Releases
+- [NPM v3.1.1] https://www.npmjs.com/package/@codat/sync-for-expenses/v/3.1.1 - sync-for-expenses